requirements.txt
# The Oracle 2026 - Production Requirements
# ==========================================

# Core
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0

# API & Web
requests>=2.28.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Telegram & Streamlit
python-telegram-bot>=20.0
streamlit>=1.28.0
plotly>=5.18.0

# Scheduler
apscheduler>=3.10.0

# Timezone
pytz>=2023.3

# Config
python-dotenv>=1.0.0
